requirements.txt
biopython>=1.83
datasets>=2.18.0
huggingface-hub>=0.21.0
pyyaml>=6.0
google-generativeai>=0.5.0
tqdm>=4.66.0
requests>=2.31.0
